Convert to using s instead of mins, start simple scheduler implementation

+++ src/main/kotlin/com/nrojiani/drone/model/deliverytime/TransitTime.kt
@@ -3,14 +3,14 @@ package com.nrojiani.drone.model.deliverytime
 /**
  * The delivery time based solely on distance - does not represent actual delivery time, which can be influenced by
  * scheduling & hours of operation.
- * @param transitTimeToDestination The expected delivery time (in minutes) from starting point (e.g., a distribution
+ * @param transitTimeToDestination The expected delivery time (in seconds) from starting point (e.g., a distribution
  *                                center) to destination.
  */
-data class TransitTime(val transitTimeToDestination: Double) {
-    val sourceToDestinationTime: Double
+data class TransitTime(val transitTimeToDestination: Long) {
+    val sourceToDestinationTime: Long
         get() = transitTimeToDestination
 
-    /** Travel time (in minutes) from source to destination and then back to source */
-    val roundTripTime: Double
+    /** Travel time (in seconds) from source to destination and then back to source */
+    val roundTripTime: Long
         get() = transitTimeToDestination * 2
 }
